feat: rename project
django-bookmarks is already taken on pypi, therefore renaming to
django-bookmark
