+++ .github/workflows/release-please.yml
@@ -18,5 +18,5 @@ jobs:
       - uses: google-github-actions/release-please-action@v3
         with:
           release-type: python
-          package-name: django-bookmarks
+          package-name: django-bookmark
           bump-minor-pre-major: true
