+++ README.md
@@ -1,30 +1,30 @@
-# Django Bookmarks
+# Django Bookmark
 
 A Django app that lets users create bookmarks to Django model instances
 
 # Installation
 
-Install `django-bookmarks` using your favourite package manager.
+Install `django-bookmark` using your favourite package manager.
 
-Add `django_bookmarks` to your Django projects `INSTALLED_APPS`:
+Add `django_bookmark` to your Django projects `INSTALLED_APPS`:
 ```
 ...
-'django_bookmarks',
+'django_bookmark',
 ...
 ```
 
-Include `django_bookmarks.urls` in your URL dispatcher:
+Include `django_bookmark.urls` in your URL dispatcher:
 ```
 ...
-path("bookmarks"/, include("django_bookmarks.urls")),
+path("", include("django_bookmark.urls")),
 ...
 ```
 
 Django Bookmarks adds two routes below where you include the app:
-* `/` is a view that lists all bookmarks of the user accessing it
+* `/bookmarks/` is a view that lists all bookmarks of the user accessing it
 * `/togglebookmark/<content_type_id>/<object_id>` is a view to add or remove a bookmark, that points to the model instance specified by the `content_type_id` and the `object_id`. The bookmark will be saved with the user id of the user accessing the view.
 
-Django Bookmarks ships the `django_bookmarks` templatetags library, which currently consists of the templatetag `toggle_bookmark`.
+Django Bookmarks ships the `django_bookmark` templatetags library, which currently consists of the templatetag `toggle_bookmark`.
 The `toggle_bookmark` templatetag takes the model instance as argument and adds a `bookmark` link that to your template. This bookmark link displayes
 a bookmark symbol that is either white, if the model instance *is not* bookmarked or yellow, if the model instance *is* bookmarked. Clicking the link
 toggles the bookmarking state of the model instance for the logged in user.
