+++ django_bookmark/admin.py
