+++ django_bookmark/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class DjangoBookmarksConfig(AppConfig):
     default_auto_field = "django.db.models.AutoField"
-    name = "django_bookmarks"
+    name = "django_bookmark"
