+++ django_bookmark/models.py
