+++ django_bookmark/templates/django_bookmark/bookmark_list.html
