+++ sample_project/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "django_bookmarks",
+    "django_bookmark",
 ]
 
 MIDDLEWARE = [
